ref(ui): Move useScrollLock to sentry/components/core
Moves useScrollLock utility and its test file from static/app/utils/ to static/app/components/core/ to better organize core UI utilities alongside other core components.
